[#588][#522] Corrected `LastTest.log` check in `build.sh`
Changes in !800 introduced a new check for script failures prior to BATS test execution to `build.sh`, which checks for the existence of `LastTest.log`. However, this new check accidentally reversed the `if` and `else` blocks, leading to erroneous failure reports. This is now fixed by swapping the code in these blocks.