Make `auto_load_octo_seed_if_needed` easier to read
Previously, GCC gave the following warning on CentOS:

```
../src/auto_load_octo_seed_if_needed.c: In function ‘auto_load_octo_seed_if_needed’:
../src/auto_load_octo_seed_if_needed.c:73:6: warning: ‘auto_load_needed’ may be used uninitialized in this function [-Wmaybe-uninitialized]
   if (!auto_load_needed) {
      ^
```

The issue was that it didn't realize the `default` branch of the switch
unconditionally returned, because it didn't realize that `status` was
never `YDB_OK` in this branch and `CLEANUP_AND_RETURN_IF_NOT_YDB_OK`
always returns when the status is not `YDB_OK`.

I could have just ignored the warning, but changing it to an explicit
return makes the code easier to read in my opinion.