[YottaDB/DB/YDB#661] Explictly set ydb_chset to M not unset
Changes in YDB#661 mean that an unset ydb_chset means that it will now
run in UTF-8 mode in systems with UTF-8 locales instead of M mode as
previously, which breaks about 50% of the tests here. It needs to be now
set explicitly to ensure that it will be M mode rather than UTF-8 mode.