[#712] Fix auto-upgrade-job pipeline failure.
The job always uses the latest version of yottadb (which will include
the new ydb_env_set). But since the whole point is to check out old
commit; it checks out an old commit of Octo that won't have the fix in
`tools/get_ydb_release.sh` that was necessary, so it will fail.