[#712] Fix another pipeline UTF-8 pipeline failure
In case of DISABLE_INSTALL=ON, if ydb_chset was `UTF-8`, we did not
correctly set the path of `_ydbocto.so` to point to the UTF-8 version of
the shared library.

This should fix multiple pipeline failures seen over this weekend (like
this one: https://gitlab.com/shabiel/YDBOcto/-/jobs/1192621115).