[#595] Add support for CREATE IF NOT EXISTS and DROP IF EXISTS
If IF NOT EXISTS is supplied for a CREATE table statement and a table exists, the result is a no-op with no errors.
In this case, error type WARN_TABLE_ALREADY_EXISTS is emitted at WARNING log severity level.

If IF EXISTS is supplied for a DROP table statement and a table does not exist, the result is a no-op with no errors.
In this case, error type WARN_TABLE_DOES_NOT_EXIST is emitted at WARNING log severity level.

This commit creates the error types WARN_TABLE_ALREADY_EXISTS and WARN_TABLE_DOES_NOT_EXIST.

This commit adds the field boolean_t if_not_exists_specified to SqlTable to store whether the corresponding CREATE
TABLE statement specified IF NOT EXISTS.
It also adds the field boolean_t if_exists_specified to SqlDropTableStatement to store whether the statement
specifies IF EXISTS.

[Miscellaneous] Fix test descriptions

The test descriptions for TDT01 and TDT02 did not match what was actually being tested.
This commit fixes those descriptions to match the tests.