[#386] [QueryGenerator] Disable TABLE.* syntax in ORDER BY if OUTER JOINs are used (due to Postgres bug)
This failure was seen in test_query_generator execution. The test failed due to a crosscheck utility failure where Octo output was correct and Postgres output was not.

make_centos was the job which failed.

Was reproducible on ubuntu as well.

TSCP16.sql has the query which failed.

Octo's behavior is correct in this case. Raised a bug report for Postgres (Tracking link -> https://www.postgresql.org/message-id/flat/17068-18d0626f1d26394d%40postgresql.org).

`QueryGenerator.m` is updated to not generate such queries i.e. queries with outer join, order by and table.*. Since such queries are not generated TSCP16 is added to test this case.