[#657] Test for Octo-AIM regression
See
https://gitlab.com/YottaDB/DBMS/YDBOcto/-/merge_requests/898#note_647885316
for detailed discussion.

The summary of the issue is that AIM currently does not cross reference
nodes that do not exist, resulting in queries that are looking for NULLs
as a cross-reference value to possibly return incorrect results.

This test is here to be part of the regression test suite to ensure that
this doesn't happen.