[#737][#733] Octo-VistA CI Job/$ydb_dist bug/RPATH
CI job is called "test-vista". It runs a new Octo-VistA test suite
against the `yottadb/octo-vehu:latest-master` docker image. This docker
image currently uses the r132 YottaDB distribution but runs the latest
Octo code that is added to the pipeline. The docker image is generated
nightly during the master branch CI run.

The tests can also be run outside of the pipeline if you have a VistA
instance with an environment file that you source before starting VistA.

To enable VistA tests independent of other Octo tests and to make the
test somewhat portable (it's impossible to make the tests completely
portable as each VistA instance differs in config and data), a new CMake
option `TEST_VISTA` has been added and this in turn asks the user for
`VISTA_ENV_FILE`. The latter is what sets up the environment for running
VistA and is run by the test. You can set up the variables interactively
using a CMake GUI, or if you pass the defines to the cmake command, you
must define both variables at the same time. The actual CTest name is
`test_vista`, corresponding to the `test_vista.bats` BATS test.

The tests do not set-up or upgrade a YottaDB VistA database; therefore,
you need to ensure that your VistA env file does everything you need,
that the databases already exist and are in operational order, and that
the VistA database already uses the target YottaDB distribution that you
are using with Octo.

Misc Changes:
- init_test in `test_helpers.bash.in` has been refactored to divorce it
from ydb_env_set, as we cannot use ydb_env_set for VistA.
- `CMakeLists.txt` now address issue #733 by letting YottaDB cmake
plugin handle finding YottaDB rather than making the user set
`ydb_dist`. This issue raised it head here and I decided to address it.
- `src/CMakeLists.txt` sets RPATH for Octo to the YottaDB distribution
path. This way we avoid needing to set LD_LIBRARY_PATH.