[YottaDB/DB/YDB#306][YottaDB/Util/YDBEncrypt#1] Fix pipeline failure due to moving encryption plugin from YDB repo to YDBEncrypt repo
Previous commits moved the encryption plugin from the YDB repo to the YDBEncrypt repo. This caused the Octo pipeline to fail because a script that installs the encryption plugin was looking for the old source.tar file instead of pulling in the YDBEncrypt repo. This commit fixes this pipeline failure by updating the install-tls script to install from the YDBEncrypt repo.