[#746] ENDPOINT clause allows a static ending point
This is like START, except it's the reverse. Unlike START, ENDPOINT
includes the ending record, if it exists.

Most of the changes are in the parsing infrastructure.

Main logic changes are in `src/m_templates/tmpl_key_end.ctemplate` and
it is tested by T039 in `tests/test_createtable.bats.in`.

`doc/grammar.rst` updated with new keyword.