[#749] IF EXISTS queries should produce INFO type messages (not WARNING)
See discussion here: YDBOctoVistA!15 (comment 684663082).

Original code that implemented IF EXISTS can be found in commits
b9140816 and 12c88029.

IF EXISTS messages should be informational, not warnings. A good
design is that they should not show up as warnings for the user since
they specifically said "I know it may not exist". Now, they don't show
up by default, which is the desired outcome.

When making this change, it was found that all IF EXISTS processing
returned a status code of 1. This is now fixed by re-doing
CLEANUP_AND_EXIT macros in `src/run_query` into two varieties:

- CLEANUP_AND_RETURN_WITH_ERROR
- CLEANUP_AND_RETURN_WITH_SUCCESS

All IF EXISTS tests have been enhanced to check the return code and to
check that locks get released.