Save ydb_gbldir env var after create_db; Create sourceable file of ydb_ env vars
Prior to this commit, the value of the `ydb_gbldir` env var that was
saved in `dbg_env.out` preceded the creation of the database, which
meant that `ydb_gbldir` pointed to the wrong database. To correct that,
`save_env_variables` has been moved from `init_test` to `create_db`.

In addition, a new custom environment export of only the `ydb` variables
using the declare syntax allows a user to cd to the test directory and
just source the new file `test.env` in order to set-up the test
environment and run yottadb and octo commands in order to examine the
data produced by the test.