[#761] DELIM "" with GLOBAL on a column crash with GVUNDEF error
It turned out that there is a small omission in Octo: columns with a
piece specification emit a `$GET` in the plans in order to protect
themselves against an undefined global; however columns without a piece
specification (e.g. using `DELIM ""` or single non-key column tables) do
not emit `$GET`.

The fix is simple, and involves a small bit of refactoring to always
emit $GET for non-extract columns, until future optimizations can be
performed where we don't need it anymore.

Note that `src/m_templates/tmpl_column_reference.ctemplate` has the
variable `skip_dollar_piece` removed. `skip_dollar_piece` was too broad
and encompassed both non-piece fields (which should have $GET) and
extract fields (which shouldn't use $GET). It's now replaced by
`is_extract`.

An assert was added to ensure that a column is not specified as a piece
and extract at the same time. This assert was triggered, leading us to
fix the cause in `src/parser/table_definition.c`. Now, EXTRACT columns
will not get a piece number assigned.

This is tested by TC050 in `tests/test_createtable.bats.in`; this query
crashes with a GVUNDEF without this change.

In addition, TVD0005 in `tests/test_vista_database.bats.in` has been
created, since this is the original query that led to the issue.
(Details on the exact failure can be found in #761.) While it will work
now on master, once
https://gitlab.com/YottaDB/DBMS/YDBOctoVistA/-/issues/7 is implemented,
this will verify that it still works with Octo.

Other ref files were updated for the following reasons:

* TC03[2-4] were updated since PIECE is not automatically assigned to
  EXTRACT fields anymore. This caused the table definition to change and
  the hash of the plan names to change, causing the ordering of the ref
  file to change.
* TCK03 it was found that tables with a single column do not emit a
  piece specification as an optimization. The composite table only has
  one column. Due to the original logic in
  `src/m_templates/tmpl_column_reference.ctemplate`, $GET was not
  emitted in the plan because there was no piece. This is now fixed.