[#768] Remove assert that causes auto-upgrade pipeline failure
As part of the changes in !1012 (merged), code was added to ensure that
a column can only be an EXTRACT or a PIECE column. However, older
versions of Octo had that uncertainty, and as a result the upgrade
pipeline failed
(https://gitlab.com/YottaDB/DBMS/YDBOcto/-/jobs/1700325926). This commit
removes the assertion until it can be reenabled again in the future.