[#776] %YDB-E-INVSTRLEN error in `find_table.c` when there are lots of objects in Octo
The cause of the issue is that macro `OCTO_SET_NULL_TERMINATED_BUFFER`
cannot accept a pointer for the second argument since it uses `sizeof`
to figure out the size, and in this case, it was sizing the buffer to
be as big as the pointer; where as we really wanted it to size it to be
as large as the array size of `oid_buff`. Passing `oid_buff` without the
pointer syntax fixes the problem.

The issue manifested itself with Octo either having an assert in Debug
mode or an error like this in Release mode if there are more than
999,999 objects (tables, columns, etc.) that were created in Octo:

```
YottaDB error: 150375522,(SimpleAPI),%YDB-E-INVSTRLEN, Invalid string length 8: max 7
```

This is tested by TC051 in `tests/test_createtable.bats.in`. The test
set the internal OID to be 999,999 and then drop and creates the NAMES
tables twice. This increments it to 1000010, which is enough to trigger
the bug. When you run a query, it runs successfully the first time, but
then fails the second time.