[YDBOctoVistA#14][#752] Computed Multiples + Infrastructure Improvements
`tests/test_vista_database.bats.in` `TVD0007` supplies a query on a computed
multiple.

This commit also takes care of #752, which removes implicit behavior for
how vista.sql is located, and makes the location of various files
explicit using the following optional CMake variables:

- TEST_VISTA_INPUT_SQL - vista.sql that is genenrated by %YDBOCTOVISTAM
- TEST_VISTA_INPUT_M   - _YDBVISTAOCTOM.m
- TEST_VISTA_INPUT_F   - _YDBVISTAOCTOF.m
- TEST_VISTA_INPUT_F_SQL - _YDBVISTAOCTOF.sql

If the CMake variables are not specified, they are defaulted to download
the files from the master branch of YDBOctoVistA repo that they
represent (with the exception of TEST_VISTA_INPUT_SQL, which is
generated from whatever version of _YDBVISTAOCTOM.m is supplied).

Also, VISTA_ENV_FILE was renamed to be TEST_VISTA_ENV_FILE, and it is
a required variable.

After this commit is merged, the pipeline in YDBOctoVistA repo needs to
be updated to use the new variables, as the pipeline won't run
successfully anymore.