[COSMETIC] Fix various tests to grep for case sensitive vs insensitive patterns in a list as appropriate
Various tests used `grep -Ei` to globally match case insensitively
against `*.m` files. This resulted in a lot of unexpected text being
included in the reference files. This caused a problem during the AIM
rebase as a lot of irrelevant content was being included.

This is now changed by making grep use the Perl Regex engine and having
it turn on and off case sensitivity when it needs to using `(?i)` and
`(?-i)`.