[#768] Reenable assertion (NULL == piece) in src/m_templates/tmpl_column_reference.ctemplate
As part of the changes in
https://gitlab.com/YottaDB/DBMS/YDBOcto/-/merge_requests/1012, code was
added to ensure that a column can only be an EXTRACT or a PIECE column.
However, older versions of Octo had that uncertainty, and as a result
the upgrade pipeline failed
(https://gitlab.com/YottaDB/DBMS/YDBOcto/-/jobs/1700325926).

The solution is to bump FMT_BINARY_DEFINITION to have the table
definition reparsed.

The original issue mentioned this as a possible solution:

> The solution is that the auto upgrade logic needs remove the
extraneous PIECE keyword if we notice an EXTRACT keyword as part of the
auto upgrade of the binary table definition, just like we did in
table_definition.c in the #761 commit.

However, it turns out that bumping FMT_BINARY_DEFINITION will reparse
the table and do the right thing; which is much less effort than the
original solution.

Note that even though the binary table definition gets recreated without
the PIECE keyword for the column that had EXTRACT specified, the text
definition of the table would stay the same and contain the extraneous
PIECE keyword. It could be confusing since the text and binary
definitions are not in sync but it does not hurt otherwise since the
binary definition is what gets used at runtime. There is no framework
that already exists to fix the text definition currently as part of the
auto upgrade logic at process startup and so it is left as is.