[VALGRIND] Fix Invalid read of size 4 in handle_parse
The logic in handle_parse allocated 8 array pieces in `handle_parse.c`,
for bind parameters, but the logic in `src/parser.y` didn't increase the
array size in case the total parameters exceeded 8. This resulted in us
trying to access invalid memory on the 8th parameter (due to the use of
`cur_parm <= parse_context.is_bind_parm_size` in the original code), as
the array is zero based, and has elements 0-7.

Valgrind error:

==78545== Invalid read of size 4
==78545==    at 0x12273C: handle_parse (handle_parse.c:256)
==78545==    by 0x127FF4: rocto_main_loop (rocto_main_loop.c:112)
==78545==    by 0x118D3A: main (rocto.c:600)
==78545==  Address 0x5b04d90 is 0 bytes after a block of size 32 alloc'd
==78545==    at 0x48435FF: calloc (vg_replace_malloc.c:1117)
==78545==    by 0x121C95: handle_parse (handle_parse.c:118)
==78545==    by 0x127FF4: rocto_main_loop (rocto_main_loop.c:112)
==78545==    by 0x118D3A: main (rocto.c:600)
==78545==

The fix is to change the `<=` to `<` in the referenced code.

Sample test that fails: TJC006.