[#205] Removed dead code in `handle_bind.c`
It was found that bugs in the code in at `copy_bind_parameter.c` line 91 and `handle_bind.c` line 656 don't cause any test failures. The reason for this is that the string stored in `bound_query` in `handle_bind.c` never gets used for anything, causing the bugs to have no observable effect.

In other words, the string in `bound_query` is simply stored on the "portal" but never looked up. This was likely included as a nice-to-have in case someone wanted to look at the query string with parameters inserted, but there is no way for clients to do this, so it is dead code that is also untested.

That said, the database node with this string can viewed in the output of some tests (e.g. `TJC010`), but it isn't validated anywhere as it plays no role in user-visible behavior. Arguably, it could be used in debugging, but no need has been found for this. Accordingly, the offending dead code has been removed from `handle_bind.c`.

This removal entailed the deletion of the `ERR_ROCTO_QUERY_TOO_LONG` code path in `copy_binary_parameter()`, since this error only came up in connection with the `bound_query`-related code removed from `handle_bind.c`. Similarly, `copy_text_parameter()` was removed from `copy_bind_parameter.c` entirely, as it is no longer called in `handle_bind.c`, and thus also removed from `rocto.h`.

These changes entailed several test updates to prevent failures in pre-existing tests:
+ Updated outref for `TMU02`
+ Revised `TJC015` and `TJC017` to validate new, correct behavior:
    - `TJC015`: A single `YDB-E-MAXSTRLEN` error is issued, but not `ERR_ROCTO_QUERY_TOO_LONG`
    - `TJC017`: no errors are issued
+ Disabled `TJC015` and `TJC017` as they run longer than the test suite timeout, since `ERR_ROCTO_QUERY_TOO_LONG` is no longer issued in these cases, which previously cut query execution to a reasonable runtime

Also `clang-tidy` now issues one fewer warning due to the offending line being deleted from `handle_bind.c`, so `clang_tidy_warnings-release.ref` and `clang_tidy_warnings_rocky-release.ref` were updated accordingly.