[ASAN] Fix Stack Buffer Overflow in generate_routine_name.c:68
This error happens on almost every single Octo operation.

```
==34977==ERROR: AddressSanitizer: stack-buffer-overflow on address 0x7ffdf6c68d1f at pc 0x557c02d25c44 bp 0x7ffdf6c67fd0 sp 0x7ffdf6c67fc0
WRITE of size 1 at 0x7ffdf6c68d1f thread T0
    #0 0x557c02d25c43 in generate_routine_name /home/sam/work/gitlab/YDBOcto/src/generate_routine_name.c:68
    #1 0x557c02dfc040 in run_query /home/sam/work/gitlab/YDBOcto/src/run_query.c:549
    #2 0x557c02cf6885 in auto_load_octo_seed /home/sam/work/gitlab/YDBOcto/src/auto_load_octo_seed.c:133
    #3 0x557c02cf7a10 in auto_load_octo_seed_if_needed /home/sam/work/gitlab/YDBOcto/src/auto_load_octo_seed_if_needed.c:100
    #4 0x557c02d9469e in octo_init /home/sam/work/gitlab/YDBOcto/src/octo_init.c:886
    #5 0x557c02cf36ef in main /home/sam/work/gitlab/YDBOcto/src/octo.c:38
    #6 0x7f1a919070b2 in __libc_start_main (/lib/x86_64-linux-gnu/libc.so.6+0x270b2)
    #7 0x557c02cf34ed in _start (/home/sam/work/gitlab/YDBOcto/build/src/octo+0x644ed)
```

The problem is across Octo, we pass a stack allocated variable to
`generate_routine_name`, but the stack allocated variable is only long
enough to be able to the actual routine name, but no NULL terminator.
`generate_routine_name` puts a null terminator in line 68, causing the
stack buffer overflow.

There are two ways of solving this:
- Remove NULL terminator from `generate_routine_name`
- Add space for NULL terminator and change various lines to substract 1 from the sizeof length.

It turns out we use the output of generate_routine_name as a `char *`,
which needs to have a null terminator. So the first solution is out of
the window, so we have to do the second solution.