[ASAN] Fix Stack Buffer Overflow in rocto.c:196
This error happens upon connection to rocto from any client.

```
==21254==ERROR: AddressSanitizer: stack-buffer-overflow on address 0x7ffd50368608 at pc 0x7f2f881ef98b bp 0x7ffd50368460 sp 0x7ffd50367c08
WRITE of size 16 at 0x7ffd50368608 thread T0
    #0 0x7f2f881ef98a in __interceptor_accept /build/gcc/src/gcc/libsanitizer/sanitizer_common/sanitizer_common_interceptors.inc:2872
    #1 0x556f8c726678 in main /home/sam/work/gitlab/YDBOcto/src/rocto.c:196
    #2 0x7f2f87697b24 in __libc_start_main (/usr/lib/libc.so.6+0x27b24)
    #3 0x556f8c4d7ccd in _start (/home/sam/work/gitlab/YDBOcto/build/src/rocto+0x6cccd)
```

The cause of this error is simple; and it's rather surprising the
program worked at all.

```
63        struct sockaddr_in *       address = NULL;
...
196               if ((cfd = accept(sfd, (struct sockaddr *)&address, &addrlen)) < 0) {
```

In accept in line 196, the second parameter is supposed to be a `struct sockaddr *`.

`address` is declared to be a pointer type already on line 63. We
correct cast it to be another pointer type, but what we cast is the
pointer to a pointer of address, `**address`, not `*address` due to the
use of the ampersand. The fix is simple: don't add the ampersand.

The usage was also found in two other lines which were fixed.