[ASAN] Fix stack buffer overflow in run_query.c:404.
This happens on `DROP TABLE`. E.g. in TII01 test.

```
==69954==ERROR: AddressSanitizer: stack-buffer-overflow on address 0x7ffce37637a0 at pc 0x55da63307243 bp 0x7ffce3762b20 sp 0x7ffce3762b10
WRITE of size 1 at 0x7ffce37637a0 thread T0
    #0 0x55da63307242 in run_query /home/sam/work/gitlab/YDBOcto/src/run_query.c:404
    #1 0x55da63419129 in main /home/sam/work/gitlab/YDBOcto/src/octo.c:197
    #2 0x7f1d671f4b24 in __libc_start_main (/usr/lib/libc.so.6+0x27b24)
    #3 0x55da631ffa9d in _start (/home/sam/work/gitlab/YDBOcto/build/src/octo+0x64a9d)

Address 0x7ffce37637a0 is located in stack of thread T0 at offset 2576 in frame
    #0 0x55da63301d3f in run_query /home/sam/work/gitlab/YDBOcto/src/run_query.c:93
```

```
383                         char                 tableGVNAME[YDB_MAX_IDENT + 1];
...
404                                         tableGVNAME[firstsub - gvname] = '\0';
```

This problem happens when we try to write a NULL terminator to the end of the
array `tableGVNAME`. The `+ 1` made me think that this is already
taken care of; however, it turns out that GVNAME is actually a global, which
includes an `^`, and that's not included in the size define YDB_MAX_IDENT.
Therefore, there is no place for a null terminator after all after the ^ is put
into the array. The correct solution is to ensure that the definition has space
for both `^` and `NULL` in addition to YDB_MAX_IDENT, i.e. `YDB_MAX_IDENT
+ 2`.