[#657] Autoupgrade AIM pipeline support
If we are doing an Autoupgrade test, it is possible that we will choose
a commit prior to AIM being introduced. This means that the old commit
will not create the `aim.dat` file with the requisite characteristics,
resulting in an error message that looks like this:

```
[ERROR] ../src/run_query.c:257 2022-01-13 14:05:15 : ERR_YOTTADB : YottaDB error:
150372986,XREFDATA+477^%YDBAIM,%YDB-E-GVSUBOFLOW, Maximum combined length of subscripts exceeded,%YDB-I-GVIS,Global variable: ^%ydbAIMDAzv1cHU8dTnPzkTFUd1656(2,"Antonio del Valle Saavedra ",5)
```

This error occurrs because if we don't create an AIM region but use AIM
in the upgraded Octo, the AIM globals end up going to the DEFAULT
region, which has an unfortunate default key size of 64.

This commit creates the AIM region if we are upgrading from an pre-AIM
commit.