[#751] Replaced `netstat` usage in `portno.m` with check on `/proc/net/tcp`
Previously, `netstat` was used in the test system to determine whether a given port was in use before spawning new `rocto` processes. However, this is not portable and `netstat` is now deprecated. Accordingly its usage is now removed and replaced a by a more portable and non-deprecated check on `/proc/net/tcp`.