[YottaDB/DB/YDB#850] [Pipeline] [Docker] Re-add missing packages
This includes pkg-config, libicu-dev, en-US locale, wget and vim
(package for xxd).

This was removed in YDB#850 in an effort to shrink down the Docker images.

These will fix the Rocky Linux pipeline failures as well as the Rocky
Linux Testing Docker file.

Also fix issues producing the Octo Docker Image.  This commit fixes this
pipeline failure:
https://gitlab.com/YottaDB/DBMS/YDBOcto/-/jobs/2129693861

The Dockerfile also has some other mostly cosmetic and correctness
changes that don't change anything operation wise.

Note that this Dockerfile produces an image which is close to being
unusable. The issues are tracked here:

https://gitlab.com/YottaDB/DBMS/YDBOcto/-/issues/677