[#814] Fix spinloop when HAVING clause of outer query uses GROUP BY in sub query (fix regression in d1914e60)
Background (Copied from #814 description)
-----------------------------------------

This is an issue that was discovered while trying to reproduce a !1079 pipeline failure. Various subtests in the `test_query_generator` bats test were seen to eventually timeout once in many runs (while executing a randomly generated query) resulting in a pipeline failure.

The cause of the failure was suspected to be an issue with !1079 causing more physical plans that are dependent on each other (see https://gitlab.com/YottaDB/DBMS/YDBOcto/-/merge_requests/1079#note_870171235 for details). And that is an issue that is parallely being investigated/fixed.

Nevertheless, a consequence of that would be the query takes little longer to run. Not that it takes for ever to run. After some investigation, found a much simpler query that hangs even in the master branch.

Below is an example. This query runs forever and spinloops returning no output.

```sql
OCTO> select n1.firstname,count(*) from names n1 group by n1.firstname having exists (select n1.firstname from names n2 group by n2.firstname);
```

The reason for the loop is in the generated M code. Line 28 uses a FOR loop with the variable `%ydboctog`. Inside the for loop, at line 30 we invoke the entryref `octoPlan2` which has its own for loop at line 55 using the same variable `%ydboctog`. This means when the `octoPlan2` invocation returns, the variable `%ydboctog` in the caller has been incorrectly modified and affects the outer FOR loop  at line 28 resulting in an indefinite loop.

```m
$ grep -En "octoPlan[12]|%ydboctog" _ydboctoPJ0ko9zyOLnBNgnPWxud7GF.m
.
.
27:    SET %ydboctog=""
28:    FOR  DO:$DATA(%ydboctocursor(cursorId,"keys",5,"","","GroupBy",%ydboctog))  SET %ydboctog=$ORDER(%ydboctocursor(cursorId,"keys",5,"","","GroupBy",%ydboctog)) QUIT:(""=%ydboctog)
30:    . SET %ydboctoexpr=$$EXISTS^%ydboctoplanhelpers(6,"octoPlan2,")
33:    . . SET %ydboctoexpr=$$mval2str^%ydboctoplanhelpers($$mvalPiece^%ydboctoplanhelpers(%ydboctog,1))
36:    . . SET %ydboctoexpr=$GET(%ydboctocursor(cursorId,"keys",5,"","","GroupBy",%ydboctog,1),0)
43:octoPlan2(cursorId)
51:    . . SET %ydboctog=%ydboctoexpr
52:    . . SET %ydboctocursor(cursorId,"keys",6,"","","GroupBy",%ydboctog)=""
54:    SET %ydboctog=""
55:    FOR  DO:$DATA(%ydboctocursor(cursorId,"keys",6,"","","GroupBy",%ydboctog))  SET %ydboctog=$ORDER(%ydboctocursor(cursorId,"keys",6,"","","GroupBy",%ydboctog)) QUIT:(""=%ydboctog)
```

One needs to do a `NEW %ydboctog` inside `octoPlan2` in order to avoid this indefinite loop.

The generated plan originally did have the `NEW %ydboctog` but it was incorrectly removed as an optimization as part of d1914e60. That needs to be reverted as clearly there is a correctness issue involved.

Fix
---
* Update `src/m_templates/tmpl_physical_plan.ctemplates` to also add `%ydboctog` to the list of NEW'd variables when any kind of grouping is present
* Update `src/physical/emit_physical_plan.c` to not include `PP_YDB_OCTO_G` in the NEW list for `octoPlan0` as it is NEW'd at each plan after this change
* TGB18 is the new subtest added to `test_group_by` to validate the fix
* Other than TGB18 rest of the test reference file changes are corrections due to the new representation of `%ydboctog`