[#500] Added support for `SELECT` without `FROM` clause when `WHERE` is included
Previously, Octo would accept `SELECT` statements that omitted both `FROM` and `WHERE`, but not work when a `WHERE` was included without `FROM`. Now, Octo will accept `SELECT` statements when `WHERE` is included but `FROM` is omitted.

This change was implemented by modifying the `query_specification` parser rule in `select.y` to include the `where_clause`, `group_by_clause`, and `having_clause` rules in its definition. These rules are now parsed and included on the `t_stmt` `SqlStatement*` variable in that rule, so that they are included during subsequent query qualification and plan construction.

To test this change, a new `TSCP20` test case was added.