[#288] Update various MySQL stuff
- For CentOS/Rocky/RHEL, MySQL is available in the repos. Use that instead of
downloading it. Dockerfile-Tests.rocky has now been updated as well, but
it currently doesn't work because there is no script to initialize
MySQL in there. For Gitlab, we rely on their MySQL service.
- For HDB01, update message so that it's correct, and move verify step
to be the last one to execute the mv command first so it's clear what is
being verified (MySQL vs Postgres).