[Pipeline] Fix Autoupgrade timeout failures
Timeout failures happened to due Postgres not getting configured in
older commits, as the test to configure postgres (`hello_psql`) was
replaced with hello_db in commit c73c3037 (discussion
https://gitlab.com/YottaDB/DBMS/YDBOcto/-/merge_requests/991#note_941578211)
. Since we checkout older commits in the autoupgrade test, the
`hello_db` test was not found.

The solution is to add `hello_psql` to the list of tests that must be run
for autoupgrades, in addition to `hello_db`. This way, the test system
will run whichever of these two it finds.

In addition, we add `netstat` (found in package `net-tools`) as the older
testing system needs it. We don't use `netstat` anymore as of commit
127d190c448a2fe1d5c06152bb4478d123defab1.