[Pipeline] Remove use of ydb_dist; not used anymore by upstream repo
As part of the refactoring done by the upstream repo
https://github.com/ChristopherEdwards/docker-vista
(https://github.com/ChristopherEdwards/docker-vista/commit/9cef27b66caa7d2f77e57feaf144782606094bdf),
done to remove a lot of old code that was true when Octo was initially
designed, $ydb_dist was removed for Octo set-up as it turns out it was
not needed. It turns out that the test here relied on it existing. This
assumption is now removed.

An incidental finding in the same file found a few days ago was that it
always thought that @TEST_VISTA_INPUT_F_SQL@ will be called "vista.sql".
I found that out when I had a file called master.sql. This is now fixed.

This fixes a bunch of pipeline failures seen today. E.g. https://gitlab.com/nars1/YDBOcto/-/jobs/2505480566.