[Pipeline] Add MySQL 8.0 service to auto-upgrade pipelines
It looks like the MySQL service was missed in commit
c73c30373d3eba164c14095345add6925998f557 because it was thought it was
not needed. However, it turns out that we may need to run MySQL tests
during upgrade testing if the test gets selected randomly, so we need to
add it.

This is an example of the failure if we don't have MySQL docker image
running:

https://gitlab.com/nars1/YDBOcto/-/jobs/2519815916

```
1..1
not ok 1 HDB01 : Can communicate and load data into Postgres & MySQL
```