[#844] Moved `libcocto.so` to `$ydb_dist/plugin`
Previously, the `libcocto.so` shared library file was installed to `$ydb_dist/plugin/o`, which is for M object files, not C object files. This change corrects this by revising `src/MakeLists.txt` to install `libcocto.so` to `$ydb_dist/plugin`, where it belongs.

To reflect this change, `ydbocto.xc` was updated to point to the new location of `libcocto.c`, and a number of supporting files referencing the old `$ydb_dist/plugin/o` location were also revised accordingly.