[#827] Replaced deprecated OpenSSL function `MD5()` with new `EVP_*` functions
OpenSSL 3.0 officially deprecates the `MD5()` API used in earlier releases, leading to compiler warnings on newer systems since this function is used in `handle_password_message.c`. To resolve this issue, a preprocessor conditional was added to `rocto.h` to account for possibility of different installed OpenSSL versions.

Further, a new `MD5()` macro to act as a drop-in replacement for the deprecated OpenSSL function of the same name on systems where older versions of OpenSSL are present. This allows for existing to code to remain unchanged while preventing the compiler warnings caused by deprecation.

To prevent compiler warnings due to type mismatches between the deprecated OpenSSL `MD5()` function and the new `MD5()` macro, typecasts were added to callers of `MD5()`.

Additionally, `#include`s of OpenSSL headers are now removed from various files, as the relevant headers are now included by `rocto.h`.