[#863] Extra parentheses gets added to generated code incorrectly
An Extract field that contains a parens in quotes throws off the parser
E.g. `TEST INTEGER EXTRACT "$$pgTableIsVisible^%ydboctopgfunctions(""("")"`.

The fix is to not count parentheses if you are inside quotes while
emitting a table column that does not contain the key.

Tested by test_createtable.bats test TC053.sql.

This is not the best fix, but is commensurate with the manual parsing
that all of `tmpl_emit_source.ctemplate` currently does. A better
implementation to parse sources should use YottaDB runtime and perform
operations using $NAME and $QSUBSCRIPT; but this is left as a future
enhancement.

This change is necessary to allow
https://gitlab.com/YottaDB/DBMS/YDBOctoVistA/-/merge_requests/29 to
proceed.