[#600] Fixed syntax highlighting bug that occurs when a failing query is run after auto-upgrade
The issue was that the `leading_spaces` C global variable is not reset after auto-upgrade operations are performed, leading to incorrect spacing calculations in `yyerror.c`. See the discussion at https://gitlab.com/YottaDB/DBMS/YDBOcto/-/merge_requests/1113#note_1018382946 for more details.

This issue is now resolved by resetting `leading_spaces` to 0 as part of the `ERASE_INPUT_BUFFER()` helper macro that runs after auto-upgrade operations.

To test this change, a new `TERR042` test case was added.

Also, brackets were added to a pre-existing `if` statement in `parse_line.c` for the sake of safety and coding style.