Revert "[VALGRIND] Add missing regfree in ^%ydboctoplanhelpers"
This reverts commit a56525775ae874b0febf571598bf28e377c70513.

We found out that implementing this causes a 2x slowdown on a simple
RegEx query (`WHERE NAME LIKE 'IBU%'`). So this commit will be discarded
as the memory leak is not big and is not important enough to justify the
slowdown.

Tracked in issue https://gitlab.com/YottaDB/DBMS/YDBOcto/-/issues/873.