[YDBOctoVistA#3] Tests for Break out Variable Pointer field into separate columns
This adds tests for YDBOctoVistA#3 to the VistA pipeline

Test `TVD0010` in `test_vista_database.bats.in`.

Tests test the following:
- Variable pointers break out correctly and only one field shows the broken out
data, not the others. This is verified manually in the reference file.
- Break outs works correctly with IS NULL
- Break outs do JOINS correctly

Independent of this, a change was made to
`tests/test_vista_database.bats.in` to not regenerate `vista.sql`
everytime it is it run if it already exists. While this may have a
potential downside (what if you want it to re-generate it?), it has
always been my intention for this to be the case, and I never fully
previously carried it out.