[#884] `CREATE FUNCTION` issues an error if function name exceeds 63 characters
Previously, function names passed to `CREATE FUNCTION` calls we accepted without a character limit, creating a discrepancy with identifier usages in other places, e.g. in column names. This discrepancy is resolved with this commit.

The fix was to:
1. Move the `ERR_IDENT_LENGTH` error check logic from the `identifier` subrule of the `column_name` rule into the `identifier` rule itself, allowing the identifier length restriction to be enforced for all identifiers.
2. Revise the subrules of the `function_definition` and `drop_function_statement` rules to use `identifier` instead of `identifier_start`
    + As part of this change, a useless subrule of `function_definition` referencing `sql_keyword` was removed

Additionally, various variables named `identifier_start`/`IDENTIFIER_START` were renamed to `identifier`/`IDENTIFIER` to reflect the change in semantics due to the aforementioned parser changes.

Also, the `constraint_name` parser rule was simplified to remove a now superfluous length check, which is now done in the `identifier` rule.

Further, it was noticed that `CREATE FUNCTION`/`DROP FUNCTION` rules did not handle all scenarios of `sql_keyword` and `PARENLESS_FUNCTION`, so subrules were added to these rules to handle these omitted cases.

A new `TCF033` test case was added to test the `ERR_IDENT_LENGTH` changes, while a new `TCF034` test case was added to test the addition of the missing `CREATE FUNCTION`/`DROP FUNCTION` subrules.