[#861] Autoupgrade Pipeline: Ninja generated files no longer show extra dots in diff
Background
----------

Previously, the `test-auto-upgrade` pipeline job would occasionally fail due to the Ninja compiler using relative paths for source files, resulting in an erroneous `..` being prefixed to source file names in error messages. This behavior is now corrected by this commit.

Code changes
------------

The main changes were to `build.sh`, which now inserts `sed` calls in `test_helpers.bash.in` to remove the `..` from test output when Octo is built with Ninja. To prevent unnecessary warnings from `git`, a call to `git reset` was also shifted up a line to ensure it is called before `git checkout -`, thus preventing dangling changes from causing a warning message.

Further, to prevent failures due to test parallelization in `test-auto-upgrade` job of `build.sh`, `build.sh` now forces `hello_*` tests to run prior to the others, to ensure that the test database is populated before any tests are run against it. Note that this enforcement is only done for `test-auto-upgrade` jobs.

Also, in `bats-tests.cmake`, `hello_db` was added as a dependency for all bats tests, with the except of the VistA tests.

Miscellaneous changes
---------------------

Added installation of `python`, `python2` and/or `python3` packages to Dockerfiles, as appropriate for each platform.